# account-service/src/main/java/com/example/account/domain/DebitRequest.java
package com.example.account.domain;

import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.NotNull;
import java.math.BigDecimal;

/**
 * Value object for debit request.
 *
 * Invariants:
 * - amount > 0
 */
public record DebitRequest(
        @NotNull(message = "Amount is required")
        @DecimalMin(value = "0.01", message = "Amount must be greater than 0")
        BigDecimal amount
) {
}