# account-service/src/main/java/com/example/account/infrastructure/dto/ErrorResponse.java
package com.example.account.infrastructure.dto;

import java.time.LocalDateTime;

/**
 * Standard error response DTO.
 */
public record ErrorResponse(
        LocalDateTime timestamp,
        int status,
        String error,
        String message,
        String path
) {
    public static ErrorResponse of(int status, String error, String message, String path) {
        return new ErrorResponse(LocalDateTime.now(), status, error, message, path);
    }
}